+++ src/main.rs
@@ -4,6 +4,7 @@
 #![feature(lazy_cell)]
 #![feature(trait_upcasting)]
 #![feature(try_blocks)]
+#![feature(type_changing_struct_update)]
 
 mod cli;
 mod collection;
@@ -18,7 +19,7 @@ use crate::{
     cli::Subcommand, collection::RequestCollection, tui::Tui,
     util::data_directory,
 };
-use anyhow::Context;
+use anyhow::{bail, Context};
 use clap::Parser;
 use std::path::PathBuf;
 use tracing_subscriber::{filter::EnvFilter, prelude::*};
@@ -43,27 +44,26 @@ struct Args {
 }
 
 #[tokio::main]
-async fn main() {
+async fn main() -> anyhow::Result<()> {
     // Global initialization
     initialize_tracing().unwrap();
     let args = Args::parse();
+    let Some(collection_path) =
+        args.collection.or_else(RequestCollection::detect_path)
+    else {
+        bail!("No collection file given and none found in current directory");
+    };
 
     // Select mode based on whether request ID(s) were given
     match args.subcommand {
         // Run the TUI
         None => {
-            let collection = RequestCollection::load(args.collection)
-                .await
-                .expect("Error loading collection");
-            Tui::start(collection)
+            Tui::start(collection_path).await;
+            Ok(())
         }
 
         // Execute one request without a TUI
-        Some(subcommand) => {
-            if let Err(err) = subcommand.execute(args.collection).await {
-                eprintln!("{err:#}");
-            }
-        }
+        Some(subcommand) => subcommand.execute(collection_path).await,
     }
 }